+++ project/plugins.sbt
@@ -8,7 +8,6 @@ addSbtPlugin("com.github.sbt" % "sbt-avro" % "3.4.2")
 addSbtPlugin("com.github.sbt" % "sbt-ci-release" % "1.5.12")
 addSbtPlugin("com.github.sbt" % "sbt-unidoc" % "0.5.0")
 addSbtPlugin("com.lightbend.paradox" % "sbt-paradox" % "0.10.3")
-addSbtPlugin("com.lightbend.paradox" % "sbt-paradox-apidoc" % "1.0.0")
 addSbtPlugin("com.lightbend.sbt" % "sbt-java-formatter" % "0.8.0")
 addSbtPlugin("com.thesamet" % "sbt-protoc" % "1.0.6")
 addSbtPlugin("com.thoughtworks.sbt-api-mappings" % "sbt-api-mappings" % "3.0.2")
