--- scio-elasticsearch/common/src/main/scala/com/spotify/scio/elasticsearch/ElasticsearchOptions.scala
@@ -1,10 +1,0 @@
-package com.spotify.scio.elasticsearch
-
-import co.elastic.clients.json.{JsonpMapper, SimpleJsonpMapper}
-import org.apache.http.HttpHost
-
-final case class ElasticsearchOptions(
-  nodes: Seq[HttpHost],
-  usernameAndPassword: Option[(String, String)] = None,
-  mapperFactory: () => JsonpMapper = () => new SimpleJsonpMapper()
-)
