+++ scio-elasticsearch/common/src/main/scala/com/spotify/scio/elasticsearch/package.scala
@@ -17,7 +17,14 @@
 
 package com.spotify.scio
 
+import co.elastic.clients.elasticsearch.core.bulk.BulkOperation
+import co.elastic.clients.json.{JsonpMapper, SimpleJsonpMapper}
+import com.spotify.scio.elasticsearch.ElasticsearchIO.{RetryConfig, WriteParam}
+import com.spotify.scio.io.ClosedTap
 import com.spotify.scio.values.SCollection
+import org.apache.beam.sdk.io.elasticsearch.ElasticsearchIO.Write.BulkExecutionException
+import org.apache.http.HttpHost
+import org.joda.time.Duration
 
 /**
  * Main package for Elasticsearch APIs. Import all.
@@ -27,6 +34,50 @@ import com.spotify.scio.values.SCollection
  * }}}
  */
 package object elasticsearch extends CoderInstances {
-  implicit def elasticsearchSCollection[T](c: SCollection[T]): ElasticsearchSCollection[T] =
-    new ElasticsearchSCollection[T](c)
+
+  final case class ElasticsearchOptions(
+    nodes: Seq[HttpHost],
+    usernameAndPassword: Option[(String, String)] = None,
+    mapperFactory: () => JsonpMapper = () => new SimpleJsonpMapper()
+  )
+
+  implicit class ElasticsearchSCollection[T](@transient private val self: SCollection[T])
+      extends AnyVal {
+
+    /**
+     * Save this SCollection into Elasticsearch.
+     *
+     * @param esOptions
+     *   Elasticsearch options
+     * @param flushInterval
+     *   delays to Elasticsearch writes for rate limiting purpose
+     * @param f
+     *   function to transform arbitrary type T to Elasticsearch `DocWriteRequest`
+     * @param numOfShards
+     *   number of parallel writes to be performed, recommended setting is the number of pipeline
+     *   workers
+     * @param errorFn
+     *   function to handle error when performing Elasticsearch bulk writes
+     */
+    def saveAsElasticsearch(
+      esOptions: ElasticsearchOptions,
+      flushInterval: Duration = WriteParam.DefaultFlushInterval,
+      numOfShards: Long = WriteParam.DefaultNumShards,
+      maxBulkRequestOperations: Int = WriteParam.DefaultMaxBulkRequestOperations,
+      maxBulkRequestBytes: Long = WriteParam.DefaultMaxBulkRequestBytes,
+      errorFn: BulkExecutionException => Unit = WriteParam.DefaultErrorFn,
+      retry: RetryConfig = WriteParam.DefaultRetryConfig
+    )(f: T => Iterable[BulkOperation]): ClosedTap[Nothing] = {
+      val param = WriteParam(
+        f,
+        errorFn,
+        flushInterval,
+        numOfShards,
+        maxBulkRequestOperations,
+        maxBulkRequestBytes,
+        retry
+      )
+      self.write(ElasticsearchIO[T](esOptions))(param)
+    }
+  }
 }
